Move files to appropriate directory
