+++ src/sparsezoo/v2/file.py
