+++ tests/sparsezoo/v2/test_directory.py
@@ -19,8 +19,8 @@ import tempfile
 
 import pytest
 
-from sparsezoo.refactor import Directory, File
-from tests.sparsezoo.refactor.test_file import _create_sample_file
+from sparsezoo.v2 import Directory, File
+from tests.sparsezoo.v2.test_file import _create_sample_file
 
 
 def _create_files_directory(files_extensions, temp_dir):
