+++ tests/sparsezoo/v2/test_model_directory.py
@@ -21,9 +21,9 @@ import numpy as np
 import pytest
 
 from sparsezoo import Zoo
-from sparsezoo.refactor.model_directory import ModelDirectory
+from sparsezoo.v2.model_directory import ModelDirectory
 from sparsezoo.requests.download import download_model_get_request
-from tests.sparsezoo.refactor.test_file import _create_sample_file
+from tests.sparsezoo.v2.test_file import _create_sample_file
 
 
 @pytest.mark.parametrize(
